requirements.txt
pydicom==2.0.0
matplotlib
pandas==1.1.3
opencv-python==4.4.0.46
pypng==0.0.20
nibabel==3.2.0
pyinstaller
